add casts between signed and unsigned

+++ src/common/logging.cc
@@ -96,9 +96,9 @@ void RayLogger_log_event(DBHandle *db,
                          double timestamp) {
   std::string timestamp_string = std::to_string(timestamp);
   int status = redisAsyncCommand(db->context, NULL, NULL, "ZADD %b %b %b",
-                                 key, key_length,
+                                 key, static_cast<size_t>(key_length),
                                  timestamp_string.data(), timestamp_string.size(),
-                                 value, value_length);
+                                 value, static_cast<size_t>(value_length));
   if ((status == REDIS_ERR) || db->context->err) {
     LOG_REDIS_DEBUG(db->context, "error while logging message to event log");
   }
